+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### 0.1.6
+* Add and enable by default correction of English indefinite articles (a / an).
+  This can be disabled using the new document setting `indefiniteArticleCleanup: false`.
+
 ### 0.1.5
 
 * Make `include` command paths relative to parent script's location